# NetInterpolationSystem

+++ Code/Majestro/Majestro_Engine/NetIdMap.h
@@ -5,16 +5,16 @@
 class NetIdMap
 {
 public:
-    Entity GetOrInvalid(uint32_t netId) const
+    Entity GetOrInvalid(uint64 netId) const
     {
         auto it = mMap.find(netId);
         return (it == mMap.end()) ? 0 : it->second;
     }
 
-    void Bind(uint32_t netId, Entity e) { mMap[netId] = e; }
-    void Unbind(uint32_t netId) { mMap.erase(netId); }
+    void Bind(uint64 netId, Entity e) { mMap[netId] = e; }
+    void Unbind(uint64 netId) { mMap.erase(netId); }
 
 private:
-    std::unordered_map<uint32_t, Entity> mMap;
+    std::unordered_map<uint64, Entity> mMap;
 };
